build-binutils.py: Remove config.cache under every subdir
Fixes: <build-binutils.py: Remove config.cache before building>

+++ build-binutils.py
@@ -129,10 +129,17 @@ def cleanup(build_folder):
     Cleanup the build directory
     :param build_folder: Build directory
     """
-    fileMakeCache = build_folder.joinpath("binutils").joinpath("config.cache").as_posix()
-    if os.path.exists(fileMakeCache):
-        os.remove(fileMakeCache)
     build_folder.mkdir(parents=True, exist_ok=True)
+    for p in os.listdir(build_folder):
+        sub = build_folder.joinpath(p)
+        if os.path.isdir(sub):
+            for sub2 in os.listdir(sub):
+                fileMakeCache = sub.joinpath(sub2).joinpath("config.cache").as_posix()
+                if os.path.exists(fileMakeCache):
+                    os.remove(fileMakeCache)
+            fileMakeCache = sub.joinpath("config.cache").as_posix()
+            if os.path.exists(fileMakeCache):
+                os.remove(fileMakeCache)
 
 
 def invoke_configure(build_folder, install_folder, root_folder, target,
